✨ [STMT-146] 최초 로그인시 회원가입 기능 구현 (#34)
* :bug: [STMT-146] 멤버 id를 int 타입에서 Long 타입으로 받도록 변경

* :sparkles: [STMT-146] 최초 로그인시 회원가입 기능 구현

+++ src/main/java/com/stumeet/server/file/application/port/in/FileUploadUseCase.java
@@ -7,7 +7,7 @@ import com.stumeet.server.file.application.port.out.FileUrl;
 
 	public interface FileUploadUseCase {
 
-	FileUrl uploadUserProfileImage(int userId, MultipartFile multipartFile);
+	FileUrl uploadUserProfileImage(Long userId, MultipartFile multipartFile);
 
-	FileUrl uploadStudyActivityImage(int studyId, List<MultipartFile> multipartFile);
+	FileUrl uploadStudyActivityImage(Long studyId, List<MultipartFile> multipartFile);
 }
